KAS-2704 uitwerking detailstab
