+++ app/components/documents/document-preview/details/details-tab.hbs
@@ -1,0 +1,82 @@
+<Auk::KeyValue @key={{t "document-name"}}>
+  {{#if this.isEditingDetails}}
+    <Textarea
+      aria-labelledby="piece-name"
+      class="auk-textarea"
+      @value={{mut @piece.name}}
+    />
+  {{else}}
+    {{@piece.name}}
+  {{/if}}
+</Auk::KeyValue>
+{{#unless @loading}}
+  {{#if (or @documentType this.isEditingDetails)}}
+    <Auk::KeyValue @key={{t "document-type"}}>
+      {{#if this.isEditingDetails}}
+        <Utils::ModelSelector
+          @modelName="document-type"
+          @allowClear={{true}}
+          @hideSearch={{true}}
+          @propertyToShow="label"
+          @sortField="priority"
+          @selectedItems={{@documentType}}
+          @selectModel={{this.changeDocumentType}}
+        />
+      {{else}}
+        {{@documentType.label}}
+      {{/if}}
+    </Auk::KeyValue>
+  {{/if}}
+  {{#if (or @piece.accessLevel this.isEditingDetails)}}
+    <Auk::KeyValue @key={{t "publicity"}}>
+      <Utils::AccessLevelColorBadge
+        @accessLevel={{@accessLevel}}
+        @editingEnabled={{this.isEditingDetails}}
+        @changeAccessLevel={{@changeAccessLevel}}
+      />
+    </Auk::KeyValue>
+  {{/if}}
+{{/unless}}
+<Auk::KeyValue @key={{capitalize (t "confidentiality")}}>
+  <Utils::ConfidentialityPill
+    @confidential={{@piece.confidential}}
+    @editingEnabled={{this.isEditingDetails}}
+  />
+</Auk::KeyValue>
+<Auk::KeyValue @key={{t "file-type"}}>
+  <Utils::Documents::FileTypePill @extension={{@piece.file.extension}} />
+</Auk::KeyValue>
+<Auk::Toolbar>
+  <Auk::Toolbar::Group @position="right">
+    {{#if this.isEditingDetails}}
+      <Auk::Toolbar::Item>
+        <Auk::ButtonLink
+          @disabled={{this.saveEditDetails.isRunning}}
+          {{on "click" this.cancelEditDetails}}
+        >
+          {{t "cancel"}}
+        </Auk::ButtonLink>
+      </Auk::Toolbar::Item>
+      <Auk::Toolbar::Item>
+
+        <Auk::Button
+          @skin="primary"
+          @loading={{this.saveEditDetails.isRunning}}
+          {{on "click" (perform this.saveEditDetails)}}
+        >
+          {{t "save"}}
+        </Auk::Button>
+      </Auk::Toolbar::Item>
+    {{else}}
+      <Auk::Toolbar::Item>
+        <Auk::Button
+          @skin="tertiary"
+          @icon="pencil"
+          {{on "click" this.openEditDetails}}
+        >
+          {{t "document-edit"}}
+        </Auk::Button>
+      </Auk::Toolbar::Item>
+    {{/if}}
+  </Auk::Toolbar::Group>
+</Auk::Toolbar>
